Rename `lexicon.py` to `base_lexicon.py`
