+++ jardin/database/clients/mysql.py
@@ -1,7 +1,7 @@
 import pymysql
 
 from jardin.database.base_client import BaseClient
-from jardin.database.lexicon import BaseLexicon
+from jardin.database.base_lexicon import BaseLexicon
 
 
 class Lexicon(BaseLexicon):
